+++ Helpers/TeamManager.cs
@@ -1,11 +1,12 @@
 ﻿using Discord.WebSocket;
+using PokeCord.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PokeCord.Data
+namespace PokeCord.Helpers
 {
     public class TeamManager
     {
