+++ Services/BadgeService.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using PokeCord.Data;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
@@ -6,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PokeCord.Data
+namespace PokeCord.Services
 {
     public class BadgeService
     {
