+++ SlashCommands/LeaderboardModule.cs
@@ -1,6 +1,6 @@
 ﻿using Discord.Interactions;
 using Microsoft.Extensions.DependencyInjection;
-using PokeCord.Data;
+using PokeCord.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
